--- src/main/java/kr/couchcoding/tennis_together/entity/CourtInfo.java
@@ -1,37 +1,0 @@
-package kr.couchcoding.tennis_together.entity;
-
-import lombok.Getter;
-
-import javax.persistence.*;
-import java.time.LocalDateTime;
-import java.time.LocalTime;
-
-@Entity
-@Getter
-public class CourtInfo {
-
-    @Id
-    @GeneratedValue
-    private Long courtNo;
-    private String roadAddr;
-    private String fee;
-    private String orgUrl;
-    private Double lat;
-    private Double lon;
-    private LocalTime startTime;
-    private LocalTime endTime;
-    private String courtContact;
-    private String adtInfo;
-    private LocalDateTime fstRegDtm;
-    private LocalDateTime lstUpdDtm;
-
-    /*
-    보류
-    @ManyToOne
-    @JoinColumn(name = "loc_sd")
-    private LocCd locSd;
-    @ManyToOne
-    @JoinColumn(name = "loc_skk")
-    private LocCd locSkk;*/
-
-}
